feat(incr13): move closet package in utils directory
