+++ cmd/shortener/main.go
@@ -11,7 +11,7 @@ import (
 	"github.com/Kenny201/go-yandex-shortener.git/internal/app/shortener"
 	"github.com/Kenny201/go-yandex-shortener.git/internal/http"
 	"github.com/Kenny201/go-yandex-shortener.git/internal/http/handler"
-	"github.com/Kenny201/go-yandex-shortener.git/internal/infra/closer"
+	"github.com/Kenny201/go-yandex-shortener.git/internal/utils/closer"
 )
 
 func main() {
